Pull `EntityIterator` to the package level
Also:
 * Suppress warnings in `Repository`.
